Recreated the scene with a list of imported objects instead of the cube
