+++ Camera.cs
@@ -3,10 +3,6 @@ using OpenTK.Graphics.OpenGL;
 using OpenTK.Input;
 
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Graphics_Homework
 {
